requirements.txt
# Backend Dependencies for Traffic Law QA System

# FastAPI and Server
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.12.5

# Qdrant Vector Database
qdrant-client==1.16.2
fastembed==0.7.4
einops==0.8.1

# Reranking
sentence-transformers==5.2.0

# OpenAI
openai==2.12.0

# PDF Processing (for law-crawler)
pypdf==6.2.0

# Document Processing (for vectorDB)
llama-index-core==0.14.10